# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/api/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/single_chat": {
      "get": {
        "tags": [
          "Single Chat"
        ],
        "description": "Получение личного чата",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Тело запроса",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Получен чат",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "type": {
                    "type": "string",
                    "example": "single"
                  },
                  "pin_message": {},
                  "name": {
                    "type": "string",
                    "example": "test"
                  },
                  "avatar": {},
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 3
                        },
                        "datetime": {
                          "type": "string",
                          "example": "2024-04-10T02:09:40.322Z"
                        },
                        "text": {
                          "type": "string",
                          "example": "Message 2"
                        },
                        "from": {
                          "type": "number",
                          "example": 10
                        },
                        "read": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  },
                  "users": {
                    "type": "object",
                    "properties": {
                      "1": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "datetime_last_activity": {
                            "type": "string",
                            "example": "2024-04-03T08:05:58.011Z"
                          },
                          "name": {
                            "type": "string",
                            "example": "test"
                          },
                          "login": {
                            "type": "string",
                            "example": "test"
                          },
                          "email": {},
                          "avatar": {},
                          "github_id": {},
                          "last_activity": {
                            "type": "number",
                            "example": 10
                          }
                        }
                      },
                      "2": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 2
                          },
                          "datetime_last_activity": {
                            "type": "string",
                            "example": "2024-04-03T08:05:58.011Z"
                          },
                          "name": {
                            "type": "string",
                            "example": "test"
                          },
                          "login": {
                            "type": "string",
                            "example": "test"
                          },
                          "email": {},
                          "avatar": {},
                          "github_id": {},
                          "last_activity": {
                            "type": "number",
                            "example": 10
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          },
          "404": {
            "description": "Сущности не существует"
          }
        }
      },
      "post": {
        "tags": [
          "Single Chat"
        ],
        "description": "Создание личного чата",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Тело запроса",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Создан чат",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "type": {
                  "type": "string",
                  "example": "single"
                },
                "pin_message": {},
                "name": {
                  "type": "string",
                  "example": "test"
                },
                "avatar": {},
                "messages": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "users": {
                  "type": "object",
                  "properties": {
                    "1": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "datetime_last_activity": {
                          "type": "string",
                          "example": "2024-04-03T08:05:58.011Z"
                        },
                        "name": {
                          "type": "string",
                          "example": "test"
                        },
                        "login": {
                          "type": "string",
                          "example": "test"
                        },
                        "email": {},
                        "avatar": {},
                        "github_id": {},
                        "last_activity": {
                          "type": "number",
                          "example": 10
                        }
                      }
                    },
                    "2": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 2
                        },
                        "datetime_last_activity": {
                          "type": "string",
                          "example": "2024-04-03T08:05:58.011Z"
                        },
                        "name": {
                          "type": "string",
                          "example": "test"
                        },
                        "login": {
                          "type": "string",
                          "example": "test"
                        },
                        "email": {},
                        "avatar": {},
                        "github_id": {},
                        "last_activity": {
                          "type": "number",
                          "example": 10
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Некорректные данные запроса",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "error.message"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          },
          "409": {
            "description": "Сущность уже существует"
          }
        }
      }
    },
...
    "/api/message": {
      "get": {
        "tags": [
          "Message"
        ],
        "description": "Получение списка сообщений",
        "responses": {
          "200": {
            "description": "Получен список сообщений",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "table": {
                    "type": "string",
                    "example": "messages"
                  },
                  "db": {
                    "type": "object",
                    "properties": {}
                  },
                  "id": {
                    "type": "number",
                    "example": 3
                  },
                  "datetime": {
                    "type": "string",
                    "example": "2024-04-10T02:09:40.322Z"
                  },
                  "text": {
                    "type": "string",
                    "example": "Message 2"
                  },
                  "from": {
                    "type": "number",
                    "example": 10
                  },
                  "read": {
                    "type": "boolean",
                    "example": true
                  },
                  "chat_id": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      },
      "post": {
        "tags": [
          "Message"
        ],
        "description": "Отправка сообщения",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Тело запроса",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "text|image|video|document|resend"
                },
                "value": {
                  "type": "string",
                  "example": "value of message"
                },
                "chat_id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Некорректные данные запроса",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "error.message"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      }
    },
    "/api/message/{id}": {
      "get": {
        "tags": [
          "Message"
        ],
        "description": "Получение сообщения по id. Зачем, а главное зачем?",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Получено сообщение",
            "schema": {
              "type": "object",
              "properties": {
                "table": {
                  "type": "string",
                  "example": "messages"
                },
                "db": {
                  "type": "object",
                  "properties": {}
                },
                "id": {
                  "type": "number",
                  "example": 1
                },
                "datetime": {
                  "type": "string",
                  "example": "2024-04-10T02:09:09.151Z"
                },
                "text": {
                  "type": "string",
                  "example": "Message 1"
                },
                "from": {
                  "type": "number",
                  "example": 1
                },
                "read": {
                  "type": "boolean",
                  "example": false
                },
                "chat_id": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      }
    },
...
  }
}